structs: move the content of csi_volumes into csi
